refactor(commands): extract bug/search tests to sibling file
